Fix PR comments. Rename and extract classes used as embark fragment parameters.
