+++ app/src/main/java/com/hedvig/app/ApplicationModule.kt
@@ -30,7 +30,7 @@ import com.hedvig.app.feature.connectpayin.ConnectPaymentViewModel
 import com.hedvig.app.feature.embark.EmbarkRepository
 import com.hedvig.app.feature.embark.EmbarkViewModel
 import com.hedvig.app.feature.embark.EmbarkViewModelImpl
-import com.hedvig.app.feature.embark.passages.TextActionSetViewModel
+import com.hedvig.app.feature.embark.passages.textactionset.TextActionSetViewModel
 import com.hedvig.app.feature.home.data.HomeRepository
 import com.hedvig.app.feature.home.service.HomeTracker
 import com.hedvig.app.feature.home.ui.HomeViewModel
