+++ app/src/main/java/com/hedvig/app/feature/embark/passages/previousinsurer/PreviousInsurerAdapter.kt
@@ -12,27 +12,28 @@ import com.hedvig.app.databinding.ExpandableBottomSheetTitleBinding
 import com.hedvig.app.databinding.PreviousInsurerItemLayoutBinding
 import com.hedvig.app.util.GenericDiffUtilItemCallback
 import com.hedvig.app.util.extensions.inflate
+import com.hedvig.app.util.extensions.view.setHapticClickListener
 import com.hedvig.app.util.extensions.viewBinding
 
 class PreviousInsurerAdapter(
-    previousInsurers: List<PreviousInsurerData.PreviousInsurer>,
+    previousInsurers: List<PreviousInsurerParameter.PreviousInsurer>,
     private val requestBuilder: RequestBuilder<PictureDrawable>,
     private val onInsurerClicked: (String) -> Unit
 ) : ListAdapter<PreviousInsurerItem, PreviousInsurerAdapter.PreviousInsurerViewHolder>(GenericDiffUtilItemCallback()) {
 
     init {
-        submitList(listOf(PreviousInsurerItem.Header(text = "Insurers")) + previousInsurers.map { it.toListItem() })
+        submitList(listOf(PreviousInsurerItem.Header) + previousInsurers.map { it.toListItem() })
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) = when (viewType) {
-        R.layout.expandable_bottom_sheet_title -> PreviousInsurerViewHolder.Header(parent.inflate(R.layout.expandable_bottom_sheet_title))
-        R.layout.previous_insurer_item_layout -> PreviousInsurerViewHolder.InsurerViewHolder(parent.inflate(R.layout.previous_insurer_item_layout))
+        R.layout.expandable_bottom_sheet_title -> PreviousInsurerViewHolder.Header(parent)
+        R.layout.previous_insurer_item_layout -> PreviousInsurerViewHolder.InsurerViewHolder(parent)
         else -> throw Error("No view type found for: $viewType")
     }
 
     override fun onBindViewHolder(holder: PreviousInsurerViewHolder, position: Int) = when (holder) {
         is PreviousInsurerViewHolder.InsurerViewHolder -> holder.bind(getItem(position) as PreviousInsurerItem.Insurer, requestBuilder, onInsurerClicked)
-        is PreviousInsurerViewHolder.Header -> holder.bind(getItem(position) as PreviousInsurerItem.Header)
+        is PreviousInsurerViewHolder.Header -> holder.bind()
     }
 
     override fun getItemViewType(position: Int) = when (getItem(position)) {
@@ -42,7 +43,7 @@ class PreviousInsurerAdapter(
 
     sealed class PreviousInsurerViewHolder(view: View) : RecyclerView.ViewHolder(view) {
 
-        class InsurerViewHolder(view: View) : PreviousInsurerViewHolder(view) {
+        class InsurerViewHolder(parent: ViewGroup) : PreviousInsurerViewHolder(parent.inflate(R.layout.previous_insurer_item_layout)) {
             private val binding by viewBinding(PreviousInsurerItemLayoutBinding::bind)
 
             fun bind(item: PreviousInsurerItem.Insurer,
@@ -54,20 +55,21 @@ class PreviousInsurerAdapter(
                     .into(binding.icon)
 
                 binding.text.text = item.name
-                binding.root.setOnClickListener {
+                binding.root.setHapticClickListener {
                     onInsurerClicked(item.name)
                 }
             }
+
         }
 
-        class Header(view: View) : PreviousInsurerViewHolder(view) {
+        class Header(parent: ViewGroup) : PreviousInsurerViewHolder(parent.inflate(R.layout.expandable_bottom_sheet_title)) {
             private val binding by viewBinding(ExpandableBottomSheetTitleBinding::bind)
 
-            fun bind(item: PreviousInsurerItem.Header) {
-                binding.title.text = item.text
+            fun bind() {
+                binding.title.setText(R.string.embark_onboarding_previous_insurer_title)
             }
         }
     }
 }
 
-private fun PreviousInsurerData.PreviousInsurer.toListItem() = PreviousInsurerItem.Insurer(name, icon)
+private fun PreviousInsurerParameter.PreviousInsurer.toListItem() = PreviousInsurerItem.Insurer(name, icon)
