+++ app/src/main/java/com/hedvig/app/feature/embark/passages/previousinsurer/PreviousInsurerItem.kt
@@ -1,9 +1,7 @@
 package com.hedvig.app.feature.embark.passages.previousinsurer
 
 sealed class PreviousInsurerItem {
-    data class Header(
-        val text: String
-    ) : PreviousInsurerItem()
+    object Header : PreviousInsurerItem()
 
     data class Insurer(
         val name: String,
