+++ app/src/main/java/com/hedvig/app/feature/embark/passages/selectaction/SelectActionFragment.kt
@@ -1,19 +1,18 @@
-package com.hedvig.app.feature.embark.passages
+package com.hedvig.app.feature.embark.passages.selectaction
 
 import android.os.Bundle
-import android.os.Parcelable
 import android.view.View
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
-import com.hedvig.android.owldroid.graphql.EmbarkStoryQuery
 import com.hedvig.app.R
 import com.hedvig.app.databinding.FragmentEmbarkSelectActionBinding
 import com.hedvig.app.feature.embark.EmbarkViewModel
+import com.hedvig.app.feature.embark.passages.MessageAdapter
+import com.hedvig.app.feature.embark.passages.animateResponse
 import com.hedvig.app.util.extensions.view.hapticClicks
 import com.hedvig.app.util.extensions.viewBinding
 import e
-import kotlinx.android.parcel.Parcelize
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.flow.onEach
@@ -25,7 +24,7 @@ class SelectActionFragment : Fragment(R.layout.fragment_embark_select_action) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val data = requireArguments().getParcelable<SelectActionPassage>(DATA)
+        val data = requireArguments().getParcelable<SelectActionParameter>(DATA)
 
         if (data == null) {
             e { "Programmer error: No DATA provided to ${this.javaClass.name}" }
@@ -34,7 +33,7 @@ class SelectActionFragment : Fragment(R.layout.fragment_embark_select_action) {
 
         binding.apply {
             messages.adapter = MessageAdapter(data.messages)
-            actions.adapter = SelectActionAdapter { selectAction: SelectAction, view: View ->
+            actions.adapter = SelectActionAdapter { selectAction: SelectActionParameter.SelectAction, view: View ->
                 view.hapticClicks()
                     .mapLatest { onActionSelected(selectAction, data, response) }
                     .onEach { model.navigateToPassage(selectAction.link) }
@@ -46,7 +45,7 @@ class SelectActionFragment : Fragment(R.layout.fragment_embark_select_action) {
         }
     }
 
-    private suspend fun onActionSelected(selectAction: SelectAction, data: SelectActionPassage, response: TextView) {
+    private suspend fun onActionSelected(selectAction: SelectActionParameter.SelectAction, data: SelectActionParameter, response: TextView) {
         selectAction.keys.zip(selectAction.values).forEach { (key, value) ->
             model.putInStore(key, value)
         }
@@ -57,7 +56,7 @@ class SelectActionFragment : Fragment(R.layout.fragment_embark_select_action) {
 
     companion object {
         private const val DATA = "DATA"
-        fun newInstance(data: SelectActionPassage) =
+        fun newInstance(data: SelectActionParameter) =
             SelectActionFragment().apply {
                 arguments = Bundle().apply {
                     putParcelable(DATA, data)
@@ -65,38 +64,3 @@ class SelectActionFragment : Fragment(R.layout.fragment_embark_select_action) {
             }
     }
 }
-
-@Parcelize
-data class SelectActionPassage(
-    val messages: List<String>,
-    val actions: List<SelectAction>,
-    val passageName: String,
-) : Parcelable {
-    companion object {
-        fun from(
-            messages: List<String>,
-            data: EmbarkStoryQuery.SelectData,
-            passageName: String,
-        ) =
-            SelectActionPassage(
-                messages,
-                data.options.map {
-                    SelectAction(
-                        it.link.fragments.embarkLinkFragment.name,
-                        it.link.fragments.embarkLinkFragment.label,
-                        it.keys,
-                        it.values
-                    )
-                },
-                passageName,
-            )
-    }
-}
-
-@Parcelize
-data class SelectAction(
-    val link: String,
-    val label: String,
-    val keys: List<String>,
-    val values: List<String>,
-) : Parcelable
