+++ app/src/main/java/com/hedvig/app/feature/embark/passages/textaction/TextFieldData.kt
@@ -1,3 +1,3 @@
-package com.hedvig.app.feature.embark.passages
+package com.hedvig.app.feature.embark.passages.textaction
 
 data class TextFieldData(val key: String?, val placeholder: String?, val mask: String?)
