+++ app/src/main/java/com/hedvig/app/feature/embark/passages/textactionset/TextActionSetFragment.kt
@@ -1,19 +1,19 @@
-package com.hedvig.app.feature.embark.passages
+package com.hedvig.app.feature.embark.passages.textactionset
 
 import android.os.Bundle
-import android.os.Parcelable
 import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import com.google.android.material.textfield.TextInputEditText
-import com.hedvig.android.owldroid.graphql.EmbarkStoryQuery
 import com.hedvig.app.R
 import com.hedvig.app.databinding.FragmentTextActionSetBinding
 import com.hedvig.app.feature.embark.EmbarkViewModel
+import com.hedvig.app.feature.embark.passages.MessageAdapter
+import com.hedvig.app.feature.embark.passages.animateResponse
+import com.hedvig.app.feature.embark.passages.textaction.TextFieldData
 import com.hedvig.app.util.extensions.view.hapticClicks
 import com.hedvig.app.util.extensions.viewBinding
 import e
-import kotlinx.android.parcel.Parcelize
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.flow.onEach
@@ -26,7 +26,7 @@ class TextActionSetFragment : Fragment(R.layout.fragment_text_action_set) {
     private val binding by viewBinding(FragmentTextActionSetBinding::bind)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        val data = requireArguments().getParcelable<TextActionSetData>(DATA)
+        val data = requireArguments().getParcelable<TextActionSetParameter>(DATA)
         if (data == null) {
             e { "Programmer error: Data is null in ${this.javaClass.name}" }
             return
@@ -51,7 +51,7 @@ class TextActionSetFragment : Fragment(R.layout.fragment_text_action_set) {
         }
     }
 
-    private suspend fun FragmentTextActionSetBinding.animateSubmission(data: TextActionSetData) {
+    private suspend fun FragmentTextActionSetBinding.animateSubmission(data: TextActionSetParameter) {
         val allInput = ""
         for ((index, key) in data.keys.withIndex()) {
             val input =
@@ -63,7 +63,7 @@ class TextActionSetFragment : Fragment(R.layout.fragment_text_action_set) {
         animateResponse(response, responseText)
     }
 
-    private fun textFieldData(data: TextActionSetData): MutableList<TextFieldData> {
+    private fun textFieldData(data: TextActionSetParameter): MutableList<TextFieldData> {
         val list = mutableListOf<TextFieldData>()
         for ((index, key) in data.keys.withIndex()) {
             list.add(TextFieldData(key, data.placeholders[index], data.mask[index]))
@@ -73,34 +73,10 @@ class TextActionSetFragment : Fragment(R.layout.fragment_text_action_set) {
 
     companion object {
         private const val DATA = "DATA"
-        fun newInstance(data: TextActionSetData) = TextActionSetFragment().apply {
+        fun newInstance(data: TextActionSetParameter) = TextActionSetFragment().apply {
             arguments = Bundle().apply {
                 putParcelable(DATA, data)
             }
         }
-    }
-}
-
-@Parcelize
-data class TextActionSetData(
-    val link: String,
-    val placeholders: List<String?>,
-    val keys: List<String?>,
-    val messages: List<String>,
-    val submitLabel: String,
-    val passageName: String,
-    val mask: List<String?>,
-) : Parcelable {
-    companion object {
-        fun from(messages: List<String>, data: EmbarkStoryQuery.TextSetData, passageName: String) =
-            TextActionSetData(
-                link = data.link.fragments.embarkLinkFragment.name,
-                placeholders = data.textActions.map { it.data?.placeholder },
-                keys = data.textActions.map { it.data?.key },
-                messages = messages,
-                submitLabel = data.link.fragments.embarkLinkFragment.label,
-                passageName = passageName,
-                mask = data.textActions.map { it.data?.mask }
-            )
     }
 }
